simulator is worked

+++ TrackAndFuel/Instrumentals/TrackerDataPortAbstract.cs
@@ -5,8 +5,8 @@ namespace TrackAndFuel.Instrumentals
 {
     abstract class TrackerDataPortAbstract
     {
-        public delegate bool CallBack(List<int> data);
-        public abstract bool Open(Dictionary<string, object> property, Action<List<int>> updateDataCallback, Action disconnectCallback);
+        public delegate bool CallBack(List<byte> data);
+        public abstract bool Open(Dictionary<string, object> property, Action<List<byte>> updateDataCallback, Action disconnectCallback);
         public abstract void Close();
         public abstract bool WriteData(byte[] data);
     }
